refactor(oxfmt): Split `oxfmtrc.rs` into 3 files (#19899)
Pure refactoring; Split large `oxfmtrc.rs` into:

- format_config.rs: struct definition for `Oxftmrc` and `FormatConfig`
- to_oxfmt_options.rs: Function to convert from `FormatConfig` to `OxfmtOptions`
- to_external_options.rs: Function to prepare external formatter options `Value`
