+++ apps/oxfmt/src/core/oxfmtrc/mod.rs
@@ -1,0 +1,10 @@
+mod format_config;
+mod to_external_options;
+mod to_oxfmt_options;
+
+pub use format_config::{
+    EndOfLineConfig, FormatConfig, OxfmtOverrideConfig, Oxfmtrc, SortImportsConfig,
+    SortPackageJsonUserConfig, SortTailwindcssConfig,
+};
+pub use to_external_options::{finalize_external_options, sync_external_options};
+pub use to_oxfmt_options::{OxfmtOptions, to_oxfmt_options};
